+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/peers/SyncSource.java
@@ -16,7 +16,7 @@ package tech.pegasys.teku.networking.eth2.peers;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.networking.p2p.peer.DisconnectReason;
-import tech.pegasys.teku.networking.p2p.reputation.DefaultReputationAdjustment;
+import tech.pegasys.teku.networking.p2p.reputation.ReputationAdjustment;
 import tech.pegasys.teku.networking.p2p.rpc.RpcResponseListener;
 import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlobSidecar;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
@@ -32,7 +32,7 @@ public interface SyncSource {
   SafeFuture<Void> requestBlobSidecarsByRange(
       UInt64 startSlot, UInt64 count, RpcResponseListener<BlobSidecar> listener);
 
-  void adjustReputation(final DefaultReputationAdjustment adjustment);
+  void adjustReputation(final ReputationAdjustment adjustment);
 
   SafeFuture<Void> disconnectCleanly(DisconnectReason reason);
 }
